@@ -12,35 +12,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.4.6",
-    "@angular/cdk": "^2.0.0-beta.12",
-    "@angular/common": "~4.4.6",
-    "@angular/compiler": "^4.4.6",
-    "@angular/core": "^4.4.6",
-    "@angular/flex-layout": "^2.0.0-beta.9",
-    "@angular/forms": "~4.4.6",
-    "@angular/http": "~4.4.6",
-    "@angular/material": "^2.0.0-beta.12",
-    "@angular/platform-browser": "~4.4.6",
-    "@angular/platform-browser-dynamic": "~4.4.6",
-    "@angular/router": "~4.4.6",
+    "@angular/animations": "5.2.0",
+    "@angular/cdk": "^5.2.0",
+    "@angular/common": "5.2.0",
+    "@angular/compiler": "5.2.0",
+    "@angular/core": "5.2.0",
+    "@angular/flex-layout": "^5.0.0-beta.14",
+    "@angular/forms": "5.2.0",
+    "@angular/http": "5.2.0",
+    "@angular/material": "^5.2.0",
+    "@angular/platform-browser": "5.2.0",
+    "@angular/platform-browser-dynamic": "5.2.0",
+    "@angular/platform-server": "5.2.0",
+    "@angular/router": "5.2.0",
     "@types/chance": "^0.7.31",
     "ajv": "^5.2.3",
     "chance": "^1.0.4",
     "core-js": "^2.5.1",
     "lodash": "^4.17.4",
     "material-design-icons": "^3.0.1",
-    "rxjs": "^5.5.0",
+    "rxjs": "5.5.2",
     "sigma": "^1.2.0",
     "treant-js": "^1.0.1",
     "ts-helpers": "^1.1.1",
-    "typescript": "2.3.4",
+    "typescript": "2.4.2",
     "typescript-collections": "^1.2.3",
     "zone.js": "^0.8.18"
   },
   "devDependencies": {
     "@angular/cli": "^1.4.9",
-    "@angular/compiler-cli": "^4.4.6",
+    "@angular/compiler-cli": "5.2.0",
     "@types/jasmine": "^2.2.30",
     "@types/lodash": "^4.14.78",
     "@types/node": "^8.0.46",
